# APIdoc.yaml
openapi: 3.0.3
info:
  title: DivingRegistrationAPI
  description: API to DivingRegistration application
  version: 1.0.0
servers:
  - url: 'https://dev-restandroid.users.info.unicaen.fr/api'
  - url: 'http://127.0.0.1:8000/api'
paths:
  /adherents:
    get:
      description: get all active divers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adherent'
    post:
      description: insert a new active diver
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdherentWithConfirmedPassNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adherent'
        '422':
          description: texts are too long or (nom,prenom) is not unique, or email is not unique, or pass is not strong enough, or not confirmed, or aptitude level is incorrect
    put:
      description: update a diver
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/AdherentWithConfirmedPass'
                - properties:
                    old_pass:
                      type: string
                      format: password
                      minLength: 8
                      example: 'MotDePasse-Pas-Securis3'
              required:
                - old_pass
                - id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Adherent'
        '404':
          description: not found
        '422':
          description: texts are too long or (nom,prenom) is not unique, or email is not unique, or pass is not strong enough, or not confirmed, or old_pass does not match current password, or aptitude level is incorrect
...
  /bateaux:
    get:
      description: get all active boats
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bateau'
    post:
      description: insert a new active boat
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BateauNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bateau'
        '422':
          description: max_personnes is less than 2 or libelle is not unique
    put:
      description: update a boat
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Bateau'
              required:
                - id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bateau'
        '404':
          description: not found
        '422':
          description: max_personnes is less than 2 or libelle is not unique
...
  /personnes:
    get:
      description: get all active persons
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Personne'
    post:
      description: insert a new active person
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PersonneWithConfirmedPassNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Personne'
        '422':
          description: texts are too long or (nom,prenom) is not unique, or email is not unique, or pass is not strong enough, or not confirmed
    put:
      description: update a person
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/PersonneWithConfirmedPass'
                - properties:
                    old_pass:
                      type: string
                      format: password
                      minLength: 8
                      example: 'Dtunpo'
              required:
                - old_pass
                - id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Personne'
        '404':
          description: not found
        '422':
          description: texts are too long or (nom,prenom) is not unique, or email is not unique, or pass is not strong enough, or not confirmed, or old_pass does not match current password
...
  /personnes/{id}:
    get:
      description: get a specific person
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Personne'
        '404':
          description: not found
    put:
      description: update a person
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/PersonneWithConfirmedPassNoId'
                - properties:
                    old_pass:
                      type: string
                      format: password
                      minLength: 8
                      example: 'Dtunpo'
              required:
                - old_pass
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Personne'
        '404':
          description: not found
        '422':
          description: texts are too long or (nom,prenom) is not unique, or email is not unique, or pass is not strong enough, or not confirmed, or old_pass does not match current password
    delete:
      description: delete a person
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Personne'
            application/x-empty:
              schema:
                type: string
                description: 'no content if person is removed'
                example: ''
        '404':
          description: not found
  /plongees:
    get:
      description: get all active dives
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
    post:
      description: insert a new active dive
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlongeeNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
        '422':
          description: site, boat, moment, aptitude level, pilot, surface security or director don't exist, date not formatted as Y-m_d, less than 2 divers allowed, max divers less than min, pilot, security or director are not allowed
    put:
      description: update a dive
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlongeeNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
        '404':
          description: not found
        '422':
          description: site, boat, moment, aptitude level, pilot, surface security or director don't exist, date not formatted as Y-m_d, less than 2 divers allowed, max divers less than min, pilot, security or director are not allowed
...
  /plongees/{id}:
    get:
      description: get a specific dive
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
        '404':
          description: not found
    put:
      description: update a dive
      parameters:
        - $ref: '#/components/parameters/PathId'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlongeeNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
        '404':
          description: not found
        '422':
          description: site, boat, moment, aptitude level, pilot, surface security or director don't exist, date not formatted as Y-m_d, less than 2 divers allowed, max divers less than min, pilot, security or director are not allowed
    delete:
      description: delete a dive
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
            application/x-empty:
              schema:
                type: string
                description: 'no content if dive is removed'
                example: ''
        '404':
          description: not found
...
components:
  parameters:
    PathId:
      name: id
      in: path
      description: ID of the requested element
      required: true
      schema:
        type: integer
  schemas:
    Actif:
      type: object
      properties:
        actif:
          type: boolean
          description: true iff it can be chosen for new dives
          example: true
    Adherent:
      allOf:
        - $ref: '#/components/schemas/Personne'
        - $ref: '#/components/schemas/AdherentBase'
        - $ref: '#/components/schemas/Actif'
    AdherentBase:
      type: object
      description: A diver registered in the association
      properties:
        license:
          type: string
          description: the diver's license
          example: '9242391985562582'
        date_certificat_medical:
          type: string
          format: date
          example: '2023-05-31'
        forfait:
          type: string
          description: the diver's remaining dives
          example: '75'
        niveau:
          type: integer
          description: the aptitude ID of the diver
          example: 9
...
    AdherentWithConfirmedPassNoId:
      allOf:
        - $ref: '#/components/schemas/AdherentBase'
        - $ref: '#/components/schemas/PersonneWithConfirmedPassNoId'
    AdherentWithConfirmedPass:
      allOf:
        - $ref: '#/components/schemas/AdherentWithConfirmedPassNoId'
        - $ref: '#/components/schemas/Id'
    BateauBase:
      type: object
      description: A boat to transport divers
      properties:
        libelle:
          type: string
          description: the boat's unique name
          example: 'La Licorne'
          maxLength: 45
        max_personnes:
          type: integer
          description: the number of people that can be transported
          example: 25
          minimum: 2
    Bateau:
      allOf:
        - $ref: '#/components/schemas/Id'
        - $ref: '#/components/schemas/BateauBase'
        - $ref: '#/components/schemas/Actif'
    BateauNoId:
      allOf:
        - $ref: '#/components/schemas/BateauBase'
        - type: object
          required:
            - libelle
            - max_personne
    Id:
      type: object
      properties:
        id:
          type: integer
          description: its unique ID
          example: 3
...
    PersonneBase:
      type: object
      description: A person registered to the application
      properties:
        nom:
          type: string
          description: the person's last name
          example: 'Dupont'
          maxLength: 45
        prenom:
          type: string
          description: the person's first name
          example: 'Dupond'
          maxLength: 45
        email:
          type: string
          format: email
          description: the person's mail address
          example: 'dupondt@iut.fr'
          maxLength: 100
    PersonneRights:
      type: object
      properties:
        directeur_de_section:
          type: boolean
          description: true iff the person is the technical director
          example: false
        secretaire:
          type: boolean
          description: true iff the person is the secretary
          example: false
        securite_de_surface:
          type: boolean
          description: true iff the person can be surface security
          example: false
        pilote:
          type: boolean
          description: true iff the person can be pilot
          example: false
    Personne:
      allOf:
        - $ref: '#/components/schemas/Id'
        - $ref: '#/components/schemas/PersonneBase'
        - $ref: '#/components/schemas/Actif'
...
    PersonneWithConfirmedPassNoId:
      allOf:
        - $ref: '#/components/schemas/PersonneBase'
        - $ref: '#/components/schemas/PersonneRights'
        - type: object
          properties:
            pass:
              type: string
              format: password
              description: the person's password
              example: 'Dtunpo-2024'
              maxLength: 60
              minLength: 8
            pass_confirmation:
              type: string
              format: password
              description: the person's password
              example: 'Dtunpo-2024'
              maxLength: 60
              minLength: 8
    PersonneWithConfirmedPass:
      allOf:
        - $ref: '#/components/schemas/PersonneWithConfirmedPassNoId'
        - $ref: '#/components/schemas/Id'
    PlongeeBase:
      type: object
      description: A dive
      properties:
        lieu:
          type: integer
          description: the ID of the site to dive at
          example: 3
        bateau:
          type: integer
          description: the ID of the boad to transport divers
          example: 1
        date:
          type: string
          format: date
          description: the date the dive will happen
          example: '2024-10-20'
        moment:
          type: integer
          description: the ID of the moment the dive will happen
          example: 1
        min_plongeurs:
          type: integer
          description: the minimum number of divers needed
          example: 4
          minimum: 2
        max_plongeurs:
          type: integer
          description: the maximum number of divers allowed. Must me greater than minimum.
          example: 4
        niveau:
          type: integer
          description: the ID of the minimum aptitude level needed for divers
          example: 10
        etat:
          type: integer
          description: the ID of the state of the dive (1 created, 2 parametrized, 3 validated, 4 done, 5 cancelled)
        securite_de_surface:
          type: integer
          description: the ID of the surface security (person)
          example: 3
        pilote:
          type: integer
          description: the ID of the pilot (person)
          example: 4
        directeur_de_plongee:
          type: integer
          description: the ID of the dive director (person)
          example: 27
    Plongee:
      allOf:
        - $ref: '#/components/schemas/Id'
        - $ref: '#/components/schemas/PlongeeBase'
        - $ref: '#/components/schemas/Actif'
    PlongeeNoId:
      allOf:
        - $ref: '#/components/schemas/PlongeeBase'
...